pyproject.toml
[project]
name = "general-local-agent"
version = "0.1.0"
description = "A ChatGPT-like local OS agent that executes tasks on your laptop"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "genai>=0.9.1",
    "google-genai>=1.46.0",
    "gradio>=5.49.1",
    "grandalf>=0.8",
    "ipython>=9.6.0",
    "langchain>=1.0.2",
    "langchain-core>=1.0.0",
    "langgraph>=1.0.1",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langsmith>=0.4.37",
    "pillow>=11.3.0",
    "playwright>=1.55.0",
    "pypdf2>=3.0.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "psutil>=5.9.0",
    "pandas>=2.0.0",
    "wmi>=1.5.1; platform_system == 'Windows'",
    "screen-brightness-control>=0.21.0; platform_system == 'Windows'",
    "numpy>=2.3.4",
]

[project.optional-dependencies]
document = ["pillow", "pypdf2", "pandas"]
windows = ["wmi", "screen-brightness-control"]
